+++ pallets/template/src/mock.rs
@@ -1,0 +1,52 @@
+// Creating mock runtime here
+
+use crate::{Module, Trait};
+use sp_core::H256;
+use frame_support::{impl_outer_origin, parameter_types, weights::Weight};
+use sp_runtime::{
+  traits::{BlakeTwo256, IdentityLookup}, testing::Header, Perbill,
+};
+
+impl_outer_origin! {
+  pub enum Origin for Test {}
+}
+
+// For testing the module, we construct most of a mock runtime. This means
+// first constructing a configuration type (`Test`) which `impl`s each of the
+// configuration traits of modules we want to use.
+#[derive(Clone, Eq, PartialEq)]
+pub struct Test;
+parameter_types! {
+  pub const BlockHashCount: u64 = 250;
+  pub const MaximumBlockWeight: Weight = 1024;
+  pub const MaximumBlockLength: u32 = 2 * 1024;
+  pub const AvailableBlockRatio: Perbill = Perbill::from_percent(75);
+}
+impl system::Trait for Test {
+  type Origin = Origin;
+  type Call = ();
+  type Index = u64;
+  type BlockNumber = u64;
+  type Hash = H256;
+  type Hashing = BlakeTwo256;
+  type AccountId = u64;
+  type Lookup = IdentityLookup<Self::AccountId>;
+  type Header = Header;
+  type Event = ();
+  type BlockHashCount = BlockHashCount;
+  type MaximumBlockWeight = MaximumBlockWeight;
+  type MaximumBlockLength = MaximumBlockLength;
+  type AvailableBlockRatio = AvailableBlockRatio;
+  type Version = ();
+  type ModuleToIndex = ();
+}
+impl Trait for Test {
+  type Event = ();
+}
+pub type TemplateModule = Module<Test>;
+
+// This function basically just builds a genesis storage key/value store according to
+// our desired mockup.
+pub fn new_test_ext() -> sp_io::TestExternalities {
+  system::GenesisConfig::default().build_storage::<Test>().unwrap().into()
+}
